% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cForwardCurve}
\alias{cForwardCurve}
\title{Constant Forward Curve}
\format{A vector with 360 elements:
\describe{
  \item{rate}{discount rate}
  ...
}}
\usage{
cForwardCurve
}
\description{
A dataset containing 2 percent continuously compounded annual interest rate
for illustration purposes.
}
\keyword{datasets}
