% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fundMap}
\alias{fundMap}
\title{Fund Map for 10 Funds}
\format{A matrix with 10 rows and 5 columns:
\describe{
  \item{index name}{name for each index}
  \item{fund number}{proportion of fund allocated to a particular index}
  ...
}}
\usage{
fundMap
}
\description{
A dataset containing a default mapping from five indices to ten different
funds.
}
\keyword{datasets}
