% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step2_ScenarioGeneration.R
\name{genFundScen}
\alias{genFundScen}
\title{Generate Fund Scenerio}
\usage{
genFundScen(fundMap, indexScen)
}
\arguments{
\item{fundMap}{A numFund-by-numIndex matrix of doubles,
mapping indices to funds.}

\item{indexScen}{A numScen-by-numStep-by-numIndex array of doubles,
index scenarios.}
}
\value{
Outputs a numScen-by-numStep-by-numFund array of doubles of
  fund scenarios.
}
\description{
Calculate numScen-by-numStep-by-numFund fund scenarios based on given index
scenarios indexScen and fund map fundMap that maps indices to funds.
}
\examples{
genFundScen(fundMap, indexScen)
}
