% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{histDates}
\alias{histDates}
\title{Historical scenario dates}
\format{A vector with 175 elements:
\describe{
  \item{date}{each observation date of the historical scenarios}
  ...
}}
\usage{
histDates
}
\description{
A dataset containing the dates at which historical returns for different
indices were observed.
}
\keyword{datasets}
