% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indexNames}
\alias{indexNames}
\title{Index names}
\format{A vector with 5 elements:
\describe{
  \item{name}{name of the index}
  ...
}}
\usage{
indexNames
}
\description{
A dataset containing names for each index.
}
\keyword{datasets}
