% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indexScen}
\alias{indexScen}
\title{5 indices for 10 scenarios over 360 months}
\format{A 3D array with dimensions 10x360x5:
\describe{
  \item{scenario}{scenario number}
  \item{month}{month since valuation date}
  \item{index number}{monthly return for a particular index in one scenario
  one month}
  ...
}}
\usage{
indexScen
}
\description{
A dataset containing a 3D array, number of scenarios (10) by
number of indices (5) by number of time steps (360), of Black-Scholes return
factors for each index in each of time step and each of scenario.
}
\keyword{datasets}
