% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mCov}
\alias{mCov}
\title{Covariance Matrix for 5 Indices}
\format{A matrix with 5 rows and 5 columns:
\describe{
  \item{index number}{number for each index}
  ...
}}
\usage{
mCov
}
\description{
A dataset containing the covariance matrix among the returns of five indices.
}
\keyword{datasets}
