% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mortTable}
\alias{mortTable}
\title{Mortality Rate for Male and Female from Ages 5 to 115}
\format{A data frame with 110 rows and 3 columns:
\describe{
  \item{age}{individual's age}
  \item{male}{mortality of a male at a particular age ranging from 5 to 115}
  \item{female}{mortality of a female at a particular age ranging
  from 5 to 115}
  ...
}}
\source{
\url{https://mort.soa.org}
}
\usage{
mortTable
}
\description{
A dataset containing the mortality rates for male and female from ages 5
to 115 (table IAM 1996 from the Society of Actuaries).
}
\keyword{datasets}
