% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swapRate}
\alias{swapRate}
\title{Swap rates across 30 years}
\format{A vector with 8 elements:
\describe{
  \item{rate}{swap rate}
  ...
}}
\source{
\url{http://www.federalreserve.gov}
}
\usage{
swapRate
}
\description{
A dataset containing US swap rates for various maturities.
}
\section{Remark}{

These swap rates were assessed on 2016-02-08
}

\keyword{datasets}
