% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VAPort}
\alias{VAPort}
\title{A Randomly Generated Pool of Variable Annuities}
\format{A data frame with 19 row and 45 columns:
\describe{
  \item{recordID}{Unique identifier of the policy}
  \item{survivorShip}{Positive weighting number}
  \item{gender}{Gender of the policyholder}
  \item{productType}{Product type}
  \item{issueDate}{Issue date}
  \item{matDate}{Maturity date}
  \item{birthDate}{Birth date of the policyholder}
  \item{currentDate}{Current date}
  \item{baseFee}{M&E (Mortality & Expense) fee}
  \item{riderFee}{Rider fee}
  \item{rollUpRate}{Roll-up rate}
  \item{gbAmt}{Guaranteed benefit}
  \item{gmwbBalance}{GMWB balance}
  \item{wbWithdrawalRate}{Guaranteed withdrawal rate}
  \item{withdrawal}{Withdrawal so far}
  \item{fundNum1}{Fund number of the 1st investment fund}
  \item{fundNum2}{Fund number of the 2nd investment fund}
  \item{fundNum3}{Fund number of the 3rd investment fund}
  \item{fundNum4}{Fund number of the 4th investment fund}
  \item{fundNum5}{Fund number of the 5th investment fund}
  \item{fundNum6}{Fund number of the 6th investment fund}
  \item{fundNum7}{Fund number of the 7th investment fund}
  \item{fundNum8}{Fund number of the 8th investment fund}
  \item{fundNum9}{Fund number of the 9th investment fund}
  \item{fundNum10}{Fund number of the 10th investment fund}
  \item{fundValue1}{Fund value of the 1st investment fund}
  \item{fundValue2}{Fund value of the 2nd investment fund}
  \item{fundValue3}{Fund value of the 3rd investment fund}
  \item{fundValue4}{Fund value of the 4th investment fund}
  \item{fundValue5}{Fund value of the 5th investment fund}
  \item{fundValue6}{Fund value of the 6th investment fund}
  \item{fundValue7}{Fund value of the 7th investment fund}
  \item{fundValue8}{Fund value of the 8th investment fund}
  \item{fundValue9}{Fund value of the 9th investment fund}
  \item{fundValue10}{Fund value of the 10th investment fund}
  \item{fundFee1}{Fund management fee of the 1st investment fund}
  \item{fundFee2}{Fund management fee of the 2nd investment fund}
  \item{fundFee3}{Fund management fee of the 3rd investment fund}
  \item{fundFee4}{Fund management fee of the 4th investment fund}
  \item{fundFee5}{Fund management fee of the 5th investment fund}
  \item{fundFee6}{Fund management fee of the 6th investment fund}
  \item{fundFee7}{Fund management fee of the 7th investment fund}
  \item{fundFee8}{Fund management fee of the 8th investment fund}
  \item{fundFee9}{Fund management fee of the 9th investment fund}
  \item{fundFee10}{Fund management fee of the 10th investment fund}
  ...
}}
\usage{
VAPort
}
\description{
A dataset containing information of the policy and the policy holder.
}
\keyword{datasets}
